hermes-brain/requirements.txt
PyYAML>=6.0
pytest>=7.0
requests>=2.31
